Web.Modernpaste.Client

Description

This module provides high-level access to the modernpaste JSON API.

In the Types module, we introduce a monad transformer stack, MPResponseT which encapsulates responses from the API and threads around the client configuration between functions for us by using ReaderT.

The simplest use of this library looks something like this:

  let req =
    createPaste (CreatePasteRequest "main = print 1234"
                 Nothing
                 (Just "Hello, Haskell!")
                 (Just "haskell")
                 Nothing)
  cfg = Modernpaste "https://modernpaste.stg.fedoraproject.org"
  runMPResponseT c req

Documentation

modernpastePost

Arguments

:: (ToJSON a, FromJSON b)

=> String
The full path of the endpoint

-> a
The data to send to the endpoint

-> MPResponseT (ModernpasteResponse b)
Data we get back from the endpoint

This is the most convenient way to call into the modernpaste API.

Given a Modernpaste config, an API path, and a JSONifyable object (anything with a ToJSON instance defined on it), we POST to the API and try to parse and return the response.

If parsing fails (e.g. we get JSON that doesn't match what we expect, or is malformed), we return Nothing. Otherwise, we return a ModernpasteResponse.

If we parse but we discover that we get an error back, we return an appropriate MPError. Otherwise, we return an MPSuccess with the decoded result.

runMPResponseT :: MPConfig -> MPResponseT a -> IO (Maybe a)

createPaste

Arguments

:: CreatePasteRequest
The paste data

-> MPResponseT (ModernpasteResponse CreatePasteResponse)
Data we get back

Create a paste.