Web.Modernpaste.Client

Documentation

modernpastePost

Arguments

:: (ToJSON a, FromJSON b)

=> Modernpaste
Modernpaste site configuration

-> String
The full path of the endpoint

-> a
The data to send to the endpoint

-> IO (Maybe (ModernpasteResponse b))
Data we get back from the endpoint

This is the most convenient way to call into the modernpaste API.

Given a Modernpaste config, an API path, and a JSONifyable object (anything with a ToJSON instance defined on it), we POST to the API and try to parse and return the response.

If parsing fails (e.g. we get JSON that doesn't match what we expect, or is malformed), we return Nothing. Otherwise, we return a ModernpasteResponse.

If we parse but we discover that we get an error back, we return an appropriate MPError. Otherwise, we return an MPSuccess with the decoded result.

createPaste

Arguments

:: Modernpaste
Modernpaste site configuration

-> CreatePasteRequest
The paste data

-> IO (Maybe (ModernpasteResponse CreatePasteResponse))
Data we get back

Create a paste.